requirements.txt
Django>=5.2,<6
djangorestframework>=3.14
django-environ>=0.11
psycopg2-binary>=2.9
